+++ packages/eslint-config/src/types/index.ts
@@ -1,4 +1,9 @@
 import type config from '@pixpilot/antfu-eslint-config';
 
+type ConfigOptionsBase = NonNullable<Parameters<ConfigFuncType>[0]>;
+
 export type ConfigFuncType = typeof config;
-export type ConfigOptions = Parameters<ConfigFuncType>[0] & { prettier?: boolean };
+export type ConfigOptions = ConfigOptionsBase & {
+  prettier?: boolean;
+  test?: ConfigOptionsBase['test'] & { relaxed?: boolean };
+};