+++ packages/eslint-config-react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @pixpilot/eslint-config-react
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @pixpilot/eslint-config@2.3.0
+
 ## 0.2.2
 
 ### Patch Changes